[refactor] : FoodInventory domain 로직과 ui 로직 분리하기
